Consolidate and idiomatic-ify
