+++ samples/star/src/commonMain/kotlin/com/slack/circuit/star/di/ApplicationContext.kt
@@ -1,0 +1,8 @@
+package com.slack.circuit.star.di
+
+import dev.zacsweers.metro.Qualifier
+
+/**
+ * Qualifier to denote a `Context` or `PlatformContext` that is specifically an Application context.
+ */
+@Qualifier annotation class ApplicationContext
