--- samples/star/src/jvmCommonMain/kotlin/com/slack/circuit/star/di/ApplicationContext.kt
@@ -1,9 +1,0 @@
-// Copyright (C) 2022 Slack Technologies, LLC
-// SPDX-License-Identifier: Apache-2.0
-package com.slack.circuit.star.di
-
-import dev.zacsweers.metro.Qualifier
-
-/** Qualifier to denote a `Context` that is specifically an Application context. */
-@Qualifier
-annotation class ApplicationContext
